Limit LED refresh rate and don't update on every cycle
Setting the color of all the keys every cycle is that the numlock is on is
unnecessary. Especially breathing is not a cheap function and doesn't change
output value for every input anyway. This waste of time in refreshing LEDs is
found to cause drag in mouse movements. This commit seeks to fix this problem.

Signed-off-by: Shriramana Sharma <samjnaa@gmail.com>